ms-api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==1.3.7
beautifulsoup4==4.12.2
chromadb==0.4.18
pydantic==2.5.0
pydantic[email]==2.5.0